requirements.txt
fastapi
uvicorn
jinja2
python-multipart
yfinance
google-generativeai
python-dotenv
email-validator
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
fpdf
alpaca-py
tradingview-ta
tradingview-screener
